refactor: domain 안에 storereservation 폴더 제거
