+++ store-reservation-service/src/test/java/com/monkey/storereservationservice/presentation/controller/StoreReservationControllerApiV1Test.java
@@ -4,8 +4,8 @@ import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.monkey.storereservationservice.application.dto.request.ReqStoreReservationPostDTOApiV1;
 import com.monkey.storereservationservice.config.StoreFeignClientMockConfig;
-import com.monkey.storereservationservice.domain.storereservation.entity.StoreReservationEntity;
-import com.monkey.storereservationservice.domain.storereservation.vo.StoreReservationStatus;
+import com.monkey.storereservationservice.domain.entity.StoreReservationEntity;
+import com.monkey.storereservationservice.domain.vo.StoreReservationStatus;
 import com.monkey.storereservationservice.infrastructure.persistence.storereservation.StoreReservationJpaRepository;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.BeforeEach;
